Added separate binary for database initialization
