+++ .gitignore
@@ -1,2 +1,3 @@
 _vendor
 /ping
+/dbsetup/dbsetup
